% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ExUtil_portfolio.R
\name{ExUtil_portfolio}
\alias{ExUtil_portfolio}
\title{S3 class ExUtil_portfolio}
\description{
Class ExUtil_portfolio is designed to describe Expected Utility portfolios with
existing direct and inverse covariance matrices. It includes the following elements:
}
\section{Slots}{
\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the sample covariance matrix of the asset returns \cr
   inv_cov_mtrx \tab the inverse of the sample covariance matrix \cr
   means \tab sample mean vector estimate for the asset returns \cr
   weights \tab portfolio weights \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab expected portfolio return \cr
   Sharpe \tab portfolio Sharpe ratio \cr
}
}

\seealso{
\link{summary.ExUtil_portfolio} summary method for the class,
\link{new_MV_portfolio_custom} class constructor, \link{validate_MV_portfolio} class validator,
\link{MV_portfolio_custom} class helper.
}
