% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDShOP-package.R
\docType{package}
\name{HDShOP-package}
\alias{HDShOP-package}
\title{A set of tools for shrinkage estimators and Expected Utility portfolios}
\description{
Package HDShOP has the following three important functions: \code{\link{EUShrinkPortfolio}},
\code{\link{CovarEstim}} and \code{\link{MeanEstim}}. EUShrinkPortfolio creates Expected Utility
portfolios using shrinkage estimation methods for portfolio weights. CovarEstim computes estimates of
covariance matrices while MeanEstim- ones for mean vectors. Every of these three functions is
supplied a name of the method to use to perform estimation.
All portfolios are stored in objects of class ExUtil_portfolio. Additionally, most portfolios have
a subclass, specific to their kind, that inherits from ExUtil_portfolio. For the later class, constructor,
validator and helper functions are available, so that custom EU portfolios may be coded by users.
}
\section{Methods}{


MeanEstim: \insertCite{BOP2019}{HDShOP}, James-Stein and Bayes-Stein estimators \insertCite{Jorion1986}{HDShOP}.

CovarEstim: \insertCite{BGP2014}{HDShOP}, \insertCite{LW2020}{HDShOP}.

EUShrinkPortfolio: \insertCite{BDOPS2021}{HDShOP}, \insertCite{BDPS2019}{HDShOP}.
}

\references{
\insertAllCited{}
}
