% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_shrink.R
\name{mean_bop19}
\alias{mean_bop19}
\title{BOP shrinkage estimator}
\usage{
mean_bop19(x, mu_0)
}
\arguments{
\item{x}{a numeric matrix. Rows represent different variables, columns- observations.}

\item{mu_0}{a numeric vector. The target vector used in the construction of the shrinkage estimator of the mean vector.}
}
\value{
a numeric vector containing the shrinkage estimation of the mean vector
}
\description{
Shrinkage estimator of the high-dimensional mean vector as suggested in Bodnar et al. (2019).
}
\examples{
n<-7e2 # number of realizations
p<-.5*n # number of assets
x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)
mm <- mean_bop19(x=x, mu_0 = rep(1,p))
}
\references{
\insertRef{BOP2019}{HDShOP}
}
