% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_shrink.R
\name{mean_js}
\alias{mean_js}
\title{James-Stein shrinkage estimator of the mean vector}
\usage{
mean_js(x, mu_0)
}
\arguments{
\item{x}{a numeric matrix. Rows represent different variables, columns- observations.}

\item{mu_0}{a numeric value. The scaling parameter of the target vector used in the
construction of the shrinkage estimator of the mean vector.}
}
\value{
a numeric vector containing the James-Stein shrinkage estimator of the mean vector
}
\description{
James-Stein shrinkage estimator of the mean vector as suggested in Jorion (1986)
}
\examples{
n<-7e2 # number of realizations
p<-.5*n # number of assets
x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)
mm <- mean_js(x=x, mu_0 = rep(1,p))
}
\references{
\insertRef{Jorion1986}{HDShOP}
}
