% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{MVShrinkPortfolio}
\alias{MVShrinkPortfolio}
\title{Shrinkage Mean Variance portfolio}
\usage{
MVShrinkPortfolio(x, gamma, type = "shrinkage", ...)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Coefficient of risk aversion.}

\item{type}{a character. The type of methods to use to construct the portfolio.}

\item{...}{arguments to pass to portfolio constructors}
}
\value{
A portfolio in the form of an object of class MeanVar_portfolio potentially with a subclass.
See \code{\link{new_MeanVar_portfolio}} for the details of the class.
}
\description{
The main function for Mean Variance (also known as Expected Utility) portfolio construction.
It is a dispatcher using methods according to argument type.
}
\details{
The sample estimator of the mean-variance portfolio weights, which results in
a traditional MV portfolio, is calculated by
\deqn{\hat w_{MV} = \frac{S^{-1} 1}{1' S^{-1} 1} + \gamma^{-1} \hat Q \bar x \quad ,}
where \eqn{S^{-1}} and \eqn{\bar x} are the inverse of the sample covariance
matrix and the sample mean vector of asset returns respectively, \eqn{\gamma}
is the coefficient of risk aversion and \eqn{\hat Q} is given by
\deqn{\hat Q = S^{-1} - \frac{S^{-1} 1 1' S^{-1}}{1' S^{-1} 1} .}
Then the shrinkage estimator for MV portfolio weights in a high-dimensional
setting is given by \deqn{\hat w_{GSMV} = \hat \alpha \hat w_{MV} + (1- \hat \alpha)b \quad,}
where \eqn{\hat \alpha} is the estimated shrinkage intensity and \eqn{b} is
a target vector.

In the case \eqn{\gamma \neq \infty}, estimation of \eqn{\alpha} is based on
a procedure proposed by Eq.2.28 of \insertCite{BOP16;textual}{HDShOP}. The case of a fully risk averse investor
(\eqn{\gamma=\infty}) leads  to the construction of the traditional global minimum
variance portfolio (GMVP):
\deqn{\hat w_{GMV} = \frac{S^{-1} 1}{1' S^{-1} 1} .}
The shrinkage estimator for the GMV portfolio is then calculated by
\deqn{\hat w_{GS GMV} = \hat \alpha \hat w_{GMV} + (1-\hat \alpha)b \quad,}
with \eqn{\hat \alpha} as in Eq.2.31 \insertCite{BPS2018;textual}{HDShOP}.

These three estimation methods are available as separate functions dispatched
accordingly to the following parameter configurations:\tabular{llll}{
   Function \tab Paper \tab Type \tab gamma \cr
   \code{\link{new_MV_portfolio_weights_BDOPS21}} \tab Bodnar et al 2021 \tab shrinkage \tab < Inf \cr
   \code{\link{new_GMV_portfolio_weights_BDPS19}} \tab Bodnar et al 2019 \tab shrinkage \tab Inf \cr
   \code{\link{new_MV_portfolio_traditional}} \tab  \tab traditional \tab > 0 \cr
}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- MVShrinkPortfolio(x=x, gamma=gamma, type='shrinkage', b=b, beta = 0.05)
str(test)

test <- MVShrinkPortfolio(x=x, gamma=Inf, type='shrinkage', b=b, beta = 0.05)
str(test)

test <- MVShrinkPortfolio(x=x, gamma=gamma, type='traditional')
str(test)

}
\references{
\insertAllCited{}
}
