% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_cov_portfol.R
\name{new_ExUtil_portfolio_cov_LW02}
\alias{new_ExUtil_portfolio_cov_LW02}
\title{constructor of EU portfolio object. Ledoit and Wolf 2020}
\usage{
new_ExUtil_portfolio_cov_LW02(x, gamma)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk aversion.}
}
\value{
an object of class ExUtil_portfolio with subclass ExUtil_portfolio_cov_LW02.\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the covariance matrix estimate of the asset returns according to Ledoit and Wolf 2020 \cr
   inv_cov_mtrx \tab the inverse of the covariance matrix estimate \cr
   means \tab sample mean vector estimate for the asset returns \cr
   W_EU_hat \tab portfolio weights estimate obtained via the above cov_mtrx and means \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab portfolio mean returns \cr
   Sharpe \tab portfolio Sharpe ratio \cr
}
}
\description{
constructor of EU portfolio object. Ledoit and Wolf 2020
}
\examples{
c <- 0.4
n <- 7e2
p <- c*n
gamma <- 1

X <- matrix(data=rnorm(n*p), nrow=p, ncol=n)
port_LW02 <- new_ExUtil_portfolio_cov_LW02(x=X, gamma)
str(port_LW02)

}
\references{
\insertRef{LW2020}{HDShOP}
}
