% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_cov_portfol.R
\name{new_ExUtil_portfolio_icov_BGP16}
\alias{new_ExUtil_portfolio_icov_BGP16}
\title{Portfolio with inverse covariance shrinkage}
\usage{
new_ExUtil_portfolio_icov_BGP16(x, gamma, TM)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk.}

\item{TM}{the target matrix for shrinkage of the inverse covariance matrix}
}
\value{
an object of class ExUtil_portfolio with a subclass ... .\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the sample covariance matrix of the assets \cr
   inv_cov_mtrx \tab the inverse of the covariance matrix estimate obtained via the method from Bodnar et al, 2016 \cr
   means \tab sample mean vector estimate for the assets \cr
   W_EU_hat \tab portfolio weights estimate obtained via the above inv_cov_mtrx and means \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab portfolio mean returns \cr
   Sharpe \tab portfolio Sharpe coefficient \cr
}
}
\description{
Portfolio with inverse covariance shrinkage
}
\examples{
c <- 0.4
n <- 7e2
p <- c*n
gamma <- 1

Sigma=matrix(0, p, p)
diag(Sigma) <- 1

X <- matrix(data=rnorm(n*p), nrow=p, ncol=n)
port_BGP16 <- new_ExUtil_portfolio_icov_BGP16(x=X, gamma, TM=Sigma)
str(port_BGP16)

}
\references{
\insertRef{BGP2016}{HDShOP}
}
