% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_shrink.R
\name{nonlin_shrinkLW}
\alias{nonlin_shrinkLW}
\title{Ledoit-Wolf Nonlinear shrinkage estimator (Annals of Statistics (2020))}
\usage{
nonlin_shrinkLW(x)
}
\arguments{
\item{x}{a numeric matrix. Rows represent different variables, columns- observations.}
}
\description{
Returns the estimation of the covariance matrix.
}
\examples{
n<-5e2
c<-0.7
p<-c*n
mu <- rep(0, p)
Sigma <- RandCovMtrx(n=n, p=p, q=20.55)

X <- t(MASS::mvrnorm(n=n, mu=mu, Sigma=Sigma))
Sigma_shr <- nonlin_shrinkLW(X)
}
\references{
\insertRef{LW2020}{HDShOP}
}
