% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_cov_matrix.R
\name{RandCovMtrx}
\alias{RandCovMtrx}
\title{Covariance matrix generator}
\usage{
RandCovMtrx(n = 300, p = 200, q = 20.55)
}
\description{
Useful for examples and tests when an arbitrary covariance is needed.
}
\examples{

n<-3e2 # number of realizations
p<-.5*n # number of assets
# A non-diagonal covariance matrix
Mtrx <- RandCovMtrx(n=n, p=p, q=20.55)
Mtrx[1:6,1:6]
}
