% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_alpha_statistics.R
\name{T_alpha}
\alias{T_alpha}
\title{T statistic for shrinkage intensity}
\usage{
T_alpha(gamma, x, w_0, beta = 0.05)
}
\arguments{
\item{gamma}{a numeric variable. Investors attitude towards risk.}

\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{w_0}{a numeric vector of tested weights.}

\item{beta}{a confidence level for alpha}
}
\value{
\tabular{ll}{
   Element \tab Description \cr
   alpha_hat \tab  \cr
   alpha_sd \tab  \cr
   alpha_lower \tab  \cr
   alpha_upper \tab  \cr
   T_alpha \tab  \cr
   p_value \tab  \cr
}
}
\description{
An asymptotic test of a given EU portfolio. The tested hypotheses are
\deqn{H_0: w_EU = w_0 \quad vs \quad H_1: w_{EU} \neq w_0.}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

T_alpha <- T_alpha(gamma=gamma, x=x, w_0=b, beta=0.05)
T_alpha
}
\references{
\insertRef{BDOPS2021}{HDShOP}
}
