% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDShOP-package.R
\docType{package}
\name{HDShOP-package}
\alias{HDShOP-package}
\title{A set of tools for shrinkage estimators and Expected Utility portfolios}
\description{
Package HDShOP has the following three important functions: \code{\link{EUShrinkPortfolio}},
\code{\link{CovarEstim}} and \code{\link{MeanEstim}}. EUShrinkPortfolio creates Expected Utility
portfolios using shrinkage estimation methods including ones for mean vectors, direct and inverse
covariance matrices and portfolio weights. CovarEstim computes sample covariance matrices while MeanEstim-
mean vectors. Every of these three functions is supplied a name of the method to use to perform estimation.
Portfolios are stored in objects of class ExUtil_portfolio. For this class constructor, validator and
helper functions are available.
}
\section{Methods}{


MeanEstim: \insertCite{BOP2019}{HDShOP}, James-Stein and Bayes-Stein estimators \insertCite{Jorion1986}{HDShOP}.

CovarEstim: \insertCite{BGP2014}{HDShOP}, \insertCite{LW2020}{HDShOP}.

EUShrinkPortfolio: all above plus \insertCite{BDOPS2021}{HDShOP} and \insertCite{BGP2016}{HDShOP}.
}

\references{
\insertAllCited{}
}
