% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{MeanVar_portfolio}
\alias{MeanVar_portfolio}
\title{A helper function for MeanVar_portfolio}
\usage{
MeanVar_portfolio(mean_vec, cov_mtrx, gamma)
}
\arguments{
\item{mean_vec}{mean vector or list of asset returns.}

\item{cov_mtrx}{the covariance matrix of asset returns. Could be a matrix or a data frame.}

\item{gamma}{a numeric variable. Coefficient of risk aversion.}
}
\description{
A helper function for MeanVar_portfolio
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

# Simple MV portfolio
cov_mtrx <- Sigma_sample_estimator(x)
means <- rowMeans(x)

cust_port_simp <- MeanVar_portfolio(mean_vec=means, cov_mtrx=cov_mtrx, gamma=2)
str(cust_port_simp)
}
