% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{MVShrinkPortfolio}
\alias{MVShrinkPortfolio}
\title{Shrinkage Expected Utility portfolio}
\usage{
MVShrinkPortfolio(x, gamma, type = "shrinkage", ...)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk aversion.}

\item{type}{a character. The type of methods to use to construct the portfolio.}

\item{...}{arguments to pass to portfolio constructors}
}
\value{
A portfolio in the form of an object of class ExUtil_portfolio potentially with a subclass.
See \code{\link{new_MV_portfolio_custom}} for the details of the class.
}
\description{
The main function for EU portfolio construction. It is a dispatcher using methods according
to argument type.
}
\details{
The available estimation methods are:\tabular{llll}{
   Function \tab Paper \tab Type \tab gamma \cr
   \code{\link{new_MV_portfolio_weights_BDOPS21}} \tab Bodnar et al 2021 \tab shrinkage \tab < Inf \cr
   \code{\link{new_GMV_portfolio_weights_BDPS19}} \tab Bodnar et al 2019 \tab shrinkage \tab Inf \cr
   \code{\link{new_MV_portfolio_traditional}} \tab  \tab traditional \tab > 0 \cr
}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- MVShrinkPortfolio(x=x, gamma=gamma, type='shrinkage', b=b, beta = 0.05)
str(test)

test <- MVShrinkPortfolio(x=x, gamma=Inf, type='shrinkage', b=b, beta = 0.05)
str(test)

test <- MVShrinkPortfolio(x=x, gamma=gamma, type='traditional')
str(test)

}
